# docs/swagger.yaml
definitions:
  req.CreateAdminBody:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  req.CreateCustomerBody:
    properties:
      delete_hash_flag:
        example: 0
        type: integer
      email:
        example: yuki@gmail.com
        type: string
      name:
        example: yuki
        type: string
      password:
        example: "123456"
        type: string
      wordpress_url:
        example: example.com
        type: string
    type: object
  req.User:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
  res.Admin:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  res.Auth:
    properties:
      token:
        type: string
    type: object
  res.Customer:
    properties:
      created_at:
        type: string
      delete_hash_flag:
        type: integer
      email:
        type: string
      facebook_token:
        type: string
      id:
        type: integer
      instagram_id:
        type: string
      instagram_name:
        type: string
      name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      wordpress_url:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/admins:
    get:
      description: 管理者ユーザーを一覧で取得します
      parameters:
      - in: query
        name: email
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: partialName
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 管理者ユーザー一覧取得
      tags:
      - Admin
  /admin/admins/i:
    get:
      description: ログイン中の管理者情報取得します
      parameters:
      - description: Admin ID
        in: query
        name: adminId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: ログイン中の管理者情報取得
      tags:
      - Admin
  /admin/customers:
    get:
      consumes:
      - application/json
      description: 全顧客を一覧で取得します
      parameters:
      - in: query
        name: email
        type: string
      - in: query
        name: facebookToken
        type: string
      - in: query
        name: isFacebookToken
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: partialName
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 顧客一覧取得
      tags:
      - Admin
  /admin/customers/{customerId}:
    get:
      description: 顧客情報を取得します
      parameters:
      - description: 顧客ID
        in: path
        name: customerId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 顧客情報取得
      tags:
      - Admin
  /admin/customers/{customerId}/posts:
    get:
      description: 顧客ごとの投稿データを一覧で取得します
      parameters:
      - description: 顧客ID
        in: path
        name: customerId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 投稿取得
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: 管理者としてログインします
      parameters:
      - description: ユーザー情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.User'
      produces:
      - application/json
      responses:
        "200":
          description: トークン
          schema:
            $ref: '#/definitions/res.Auth'
      summary: ログイン
      tags:
      - Admin
  /admin/register/admin:
    post:
      consumes:
      - application/json
      description: 管理者ユーザーを作成します
      parameters:
      - description: 管理者情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.CreateAdminBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Admin'
      security:
      - Token: []
      summary: 管理者ユーザーの作成
      tags:
      - Admin
  /admin/register/customer:
    post:
      consumes:
      - application/json
      description: 顧客を作成します
      parameters:
      - description: 顧客作成用リクエスト
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.CreateCustomerBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Customer'
      security:
      - Token: []
      summary: 顧客の作成
      tags:
      - Admin
  /badge/execute:
    get:
      description: 全顧客に対して、InstagramからWordpressへのデータ連携を行います。
      responses: {}
      security:
      - Token: []
      summary: バッチ実行
      tags:
      - Badge
  /customer/i:
    get:
      consumes:
      - application/json
      description: 自分の顧客情報を取得します
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 顧客情報の取得
      tags:
      - Customer
  /customer/i/fetch/post:
    post:
      description: インスタグラムとWordpressの連携
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: インスタグラムとWordpressの連携
      tags:
      - Customer
  /customer/i/posts:
    get:
      consumes:
      - application/json
      description: 顧客ごとの投稿を一覧で取得します
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 顧客の投稿一覧の取得
      tags:
      - Customer
  /customer/login:
    post:
      consumes:
      - application/json
      description: 顧客としてログインします
      parameters:
      - description: ユーザー情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Auth'
      summary: ログイン
      tags:
      - Customer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"